Refactoring blog entries into separate directories
