+++ src/routes/my-story/blog-page/blog-page.styles.scss
@@ -1,0 +1,36 @@
+.row {
+  .left-column {
+    float: left;
+    width: 70%;
+  }
+  .right-column {
+    float: left;
+    width: 30%;
+    padding-left: 20px;
+  }
+
+  &:after {
+    content: "";
+    display: table;
+    clear: both;
+  }
+}
+
+@media screen and (max-width: 600px) {
+  .row {
+    display: flex;
+    flex-flow: column;
+
+    .left-column {
+      order: 2;
+      width: 100%;
+      padding: 0;
+    }
+
+    .right-column {
+      order: 1;
+      width: 100%;
+      padding: 0;
+    }
+  }
+}
