+++ src/routes/my-story/footer/footer.component.jsx
@@ -1,0 +1,13 @@
+import "./footer.styles.scss"
+
+function Footer() {
+  return (
+    <div className="footer">
+      <a href="https://github.com/chrisfishbob/chris-site" target="_blank">
+        <p> Chris Hsu 2022 &nbsp;&nbsp;&nbsp; Made with ReactJS</p>
+      </a>
+    </div>
+  );
+}
+
+export default Footer;
