+++ src/routes/my-story/footer/footer.styles.scss
@@ -1,0 +1,10 @@
+.footer {
+  padding: 3px;
+  text-align: center;
+  background-color: rgb(240, 240, 240);
+  margin-top: 20px;
+
+  p {
+    font-size: 15px;
+  }
+}
