+++ src/routes/my-story/left-column-items/bowling-blog/bowling-blog.styles.scss
@@ -1,0 +1,15 @@
+.bowling-blog-container {
+  background-color: rgb(240, 240, 240);
+  padding: 20px;
+  margin-top: 20px;
+
+  .bowling-img {
+    width: 70%;
+  }
+
+  @media screen and (max-width: 600px) {
+    & {
+        text-align: center;
+    }
+  }
+}
