+++ src/routes/my-story/left-column-items/deloitte-blog/deloitte-blog.styles.scss
@@ -1,0 +1,13 @@
+.deloitte-blog-container {
+  background-color: rgb(240, 240, 240);
+  padding: 20px;
+  margin-top: 20px;
+
+  .deloitte-img {
+    width: 70%;
+  }
+
+  @media screen and (max-width: 600px) {
+    text-align: center;
+  }
+}
