+++ src/routes/my-story/left-column-items/spain-blog/spain-blog.component.jsx
@@ -1,11 +1,13 @@
+import "./spain-blog.styles.scss"
+
 function SpainBlog() {
   const spain = "https://d34316robnoe1k.cloudfront.net/spain-2.jpg";
 
   return (
-    <div className="event">
+    <div className="spain-blog-container">
       <h1> Spain</h1>
       <h3> Summer, 2022</h3>
-      <img className="poly-img" src={spain} />
+      <img className="spain-img" src={spain} />
       <p>
         Got to visit Spain briefly for a family vacation! (I look different
         because of the lighting, It's not a drawing).
