--- src/routes/my-story/my-story.styles.scss
@@ -1,81 +1,0 @@
-.left-column {
-  float: left;
-  width: 70%;
-}
-
-.right-column {
-  float: left;
-  width: 30%;
-  padding-left: 20px;
-}
-
-.event, .links ,.site-info{
-  background-color: rgb(240, 240, 240);
-  padding: 20px;
-  margin-top: 20px;
-}
-
-.poly-img {
-    width: 70%;
-}
-
-.links a{
-    font-size: 20px;
-}
-
-.row:after {
-  content: "";
-  display: table;
-  clear: both;
-}
-
-.site-info {
-  .exp {
-    display: none;
-  }
-  dd:hover {
-    color: red;
-    & > .exp {
-      display: block;
-      color: red;
-    }
-  }
-}
-
-.footer {
-  padding: 3px;
-  text-align: center;
-  background-color: rgb(240, 240, 240);
-  margin-top: 20px;
-
-  p {
-    font-size: 15px;
-  }
-}
-
-@media screen and (max-width: 600px) {
-  .row {
-    display: flex;
-    flex-flow: column;
-  }
-
-  .links ,.site-info {
-    display: none;
-  }
-
-  .left-column{
-    order: 2;
-    width: 100%;
-    padding: 0;
-  }
-
-  .right-column{
-    order: 1;
-    width: 100%;
-    padding: 0;
-  }
-
-  .event {
-    text-align: center;
-  }
-}
