+++ src/routes/my-story/right-column-items/about-me/about-me.component.jsx
@@ -1,10 +1,12 @@
+import "./about-me.styles.scss"
+
 function AboutMeBlog() {
   const me = "https://d34316robnoe1k.cloudfront.net/meinspain.JPG";
 
   return (
-    <div className="event">
+    <div className="about-me-container">
       <h1> Hi There! </h1>
-      <img className="poly-img" src={me} />
+      <img className="profile-pic" src={me} />
       <p> I'm Chris and I'm a Computer Science student at Cal Poly SLO! </p>
       <p>
         I love all things computing, but I'm particularly interested in cloud
