+++ src/routes/my-story/right-column-items/about-me/about-me.styles.scss
@@ -1,0 +1,15 @@
+.about-me-container {
+  background-color: rgb(240, 240, 240);
+  padding: 20px;
+  margin-top: 20px;
+
+  .profile-pic {
+    width: 70%;
+  }
+
+  @media screen and (max-width: 600px) {
+    text-align: center;
+  }
+}
+
+
